Added new seeds for CCG and GpPractice

+++ Mep.Api/ViewModels/Patient.cs
@@ -1,4 +1,5 @@
 using System.ComponentModel.DataAnnotations;
+using System;
 
 namespace Mep.Api.ViewModels
 {
@@ -10,7 +11,7 @@ namespace Mep.Api.ViewModels
     public int? CcgId { get; set; }
     public virtual GpPractice GpPractice { get; set; }
     public int? GpPracticeId { get; set; }
-    public int? NhsNumber { get; set; }
+    public Int64? NhsNumber { get; set; }
     [MaxLength(10)]
     [Required]
     public string ResidentialPostcode { get; set; }
