+++ Mep.Data/Entities/Audit/PatientAudit.cs
@@ -1,5 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
+using System;
 
 namespace Mep.Data.Entities
 {
@@ -12,7 +13,7 @@ namespace Mep.Data.Entities
     public int? CcgId { get; set; }
     // public virtual GpPracticeAudit GpPractice { get; set; }
     public int? GpPracticeId { get; set; }
-    public int? NhsNumber { get; set; }
+    public Int64? NhsNumber { get; set; }
     [MaxLength(10)]
     [Required]
     public string ResidentialPostcode { get; set; }
